firmware: qcom_scm: Avoid integer overflow in tempory variable
Valid values of a vmid can be between 0 to 63. Define next_vm as
a u64 to avoid overflow from below.

next_vm |= BIT(newvm->vmid);

+++ drivers/firmware/qcom/qcom_scm.c
@@ -1317,7 +1317,7 @@ int qcom_scm_assign_mem(phys_addr_t mem_addr, size_t mem_sz,
 	size_t dest_sz;
 	size_t src_sz;
 	size_t ptr_sz;
-	int next_vm;
+	u64 next_vm;
 	__le32 *src;
 	int ret, i, b;
 	u64 srcvm_bits = *srcvm;
